--- src/components/SearchBar.vue
@@ -1,11 +1,0 @@
-<script setup>
-import TheIcons from './TheIcons.vue'
-</script>
-
-<template>
-  <div class="flex border border-stone-500 rounded-lg p-2">
-    <slot></slot>
-  </div>
-</template>
-
-<style scoped></style>
