+++ src/components/ansehen/TableView.vue
@@ -1,13 +1,13 @@
 <script setup>
 import { testHardware } from '@/config/testHardware'
-import SearchBar from './SearchBar.vue'
-import TheIcons from './TheIcons.vue'
+import SearchBar from '@/components/ansehen/SearchBar.vue'
+import TheIcons from '@/components/TheIcons.vue'
 import { useCardStore } from '@/stores/card'
 import { useTableStore } from '@/stores/table'
 import { useKomponenteStore } from '@/stores/komponenten'
 import { computed, ref } from 'vue'
 import { useModalStore } from '@/stores/modal'
-import { useFilterStore } from '@/stores/filterItems'
+import { useFilterStore } from '@/stores/filter'
 
 // stores
 const cardStore = useCardStore()
@@ -20,10 +20,7 @@ const modalStore = useModalStore()
 
 const filterStore = useFilterStore()
 
-// ref vars
-const filter = ref(false)
-
-// filter second table
+// filter für Komponente Table
 const filteredKomponentenZurHardware = computed(() => {
   const filteredKomponentenZurHardware = tableStore.komponentenZurHardware.filter((hardware) => {
     for (let index = 0; index < komponenteStore.komponenten.length; index++) {
@@ -37,8 +34,9 @@ const filteredKomponentenZurHardware = computed(() => {
   return filteredKomponentenZurHardware
 })
 
-// filter normal table
+// filter für normale Table
 const filteredTestHardware = computed(() => {
+  // filter anhand Treeview toggle btns
   const filteredTestHardware = testHardware.filter((hardware) => {
     for (let index = 0; index < komponenteStore.komponenten.length; index++) {
       for (let i = 0; i < komponenteStore.komponenten[index].children.length; i++) {
@@ -49,8 +47,9 @@ const filteredTestHardware = computed(() => {
     }
   })
 
-  if (filter.value === false) return filteredTestHardware
+  if (filterStore.filter === false) return filteredTestHardware
 
+  // filter anhand SearchBar inputText
   const searchHardware = testHardware.filter((hardware) => {
     const filterItem = filterStore.mapFilterItems[filterStore.currentToggleFilterItem]
     const result = hardware[filterItem].includes(filterStore.currentFilterItem[filterItem])
@@ -60,77 +59,45 @@ const filteredTestHardware = computed(() => {
   return searchHardware
 })
 
-function refreshData() {
-  tableStore.showTable = 'normalTable'
-  cardStore.cardState = false
-  cardStore.currentHardware = ''
-}
-
-function deleteFilter() {
-  filter.value = false
-  filterStore.searchBarText = ''
-}
-
+// func für SearchBtn: SearchBar inputText in reactive Object speichern und filter auf true setzen um "filter anhand SearchBar inputText" zu triggern
 function updateFilterObj() {
   const filterItem = filterStore.mapFilterItems[filterStore.currentToggleFilterItem]
 
   filterStore.currentFilterItem[filterItem] = filterStore.searchBarText
 
-  filter.value = true
+  filterStore.filter = true
 }
 
-function getFilterToggleItem() {
-  return filterStore.mapFilterItems[filterStore.currentToggleFilterItem]
+// func für SearchClearBtn: Searchbar filterState auf false setzen (= Nicht mehr auf eine Hardw. filtern), Searchbar Text leer setzen
+function deleteFilter() {
+  filterStore.filter = false
+  filterStore.searchBarText = ''
 }
 
-const showSearchBarDropdown = ref(false)
+// func für RefreshBtn: normale Tabelle anzeigen, InformationCard schließen, ausgewählte Hardware leer setzen
+function refreshData() {
+  tableStore.showTable = 'normalTable'
+  cardStore.cardState = false
+  cardStore.currentHardware = ''
+}
 </script>
 
 <template>
   <div class="flex flex-col gap-5">
+    <!-- Searchbar, SearchIcon, SearchItem, SearchClear, FilterIcon, RefreshIcon -->
     <div class="w-full flex items-center gap-2">
-      <SearchBar class="w-full h-full">
-        <TheIcons
-          @click="showSearchBarDropdown = !showSearchBarDropdown"
-          icon="bi bi-body-text"
-          class="text-2xl px-1 text-stone-600 border rounded-md"
-        ></TheIcons>
-        <input
-          type="text"
-          placeholder="Suche nach ..."
-          v-model="filterStore.searchBarText"
-          @change="updateFilterObj()"
-          class="outline-none w-full h-full ps-2"
-          :class="{ '': showSearchBarDropdown === true }"
-        />
-        <TheIcons
-          v-if="filterStore.searchBarText !== ''"
-          @click="((filterStore.searchBarText = ''), (filter = false))"
-          icon="bi bi-x"
-          class="text-2xl ps-2 text-stone-600"
-        ></TheIcons>
-        <div
-          v-if="showSearchBarDropdown"
-          class="absolute -translate-x-2 w-364 h-50 overflow-y-scroll translate-y-11 bg-white border rounded-md select-none border-stone-500"
-        >
-          <div
-            v-for="hardware in testHardware"
-            class="p-2 rounded-md cursor-pointer hover:bg-stone-200"
-            @click="
-              (((filterStore.searchBarText = hardware[getFilterToggleItem()]),
-              (showSearchBarDropdown = false)),
-              updateFilterObj())
-            "
-          >
-            {{ hardware[getFilterToggleItem()] }}
-          </div>
-        </div>
-      </SearchBar>
+      <!-- Searchbar -->
+
+      <SearchBar class="w-full h-full"> </SearchBar>
+      <!-- SearchIcon -->
+
       <TheIcons
         @click="updateFilterObj()"
         icon="bi bi-search"
         class="text-2xl border border-stone-500 p-2 h-full px-3 rounded-lg text-stone-600 hover:bg-stone-300 hover:text-white hover:duration-200 not-focus:duration-200"
       ></TheIcons>
+      <!--  SearchItem -->
+
       <div
         @click="filterStore.filterToggleBarState = !filterStore.filterToggleBarState"
         class="select-none flex gap-2 items-center text-md border border-stone-500 p-2 h-full px-3 rounded-lg text-stone-600 hover:bg-stone-200 hover:text-stone-800 hover:duration-300 not-focus:duration-200"
@@ -138,10 +105,12 @@ const showSearchBarDropdown = ref(false)
         {{ filterStore.currentToggleFilterItem }}
         <TheIcons v-if="filterStore.filterToggleBarState === false" icon="bi bi-caret-down" />
         <TheIcons v-if="filterStore.filterToggleBarState" icon="bi bi-caret-down-fill" />
+        <!-- SearchItem: ToggleBarDown -->
         <div
           v-if="filterStore.filterToggleBarState"
           class="absolute -translate-x-3 translate-y-42 bg-white border rounded-md border-stone-500 text-center"
         >
+          <!-- ToggleBarDown -> Items Iteration -->
           <div
             v-for="item in filterStore.filterItemsCapStr"
             class="p-2 rounded-md hover:bg-stone-200"
@@ -151,6 +120,8 @@ const showSearchBarDropdown = ref(false)
           </div>
         </div>
       </div>
+      <!-- SearchClear -->
+
       <div
         @click="deleteFilter()"
         class="uppercase select-none flex gap-2 items-center text-md p-2 h-full px-3 rounded-lg text-stone-600 hover:bg-stone-100 hover:duration-200 not-focus:duration-200"
@@ -158,17 +129,24 @@ const showSearchBarDropdown = ref(false)
         Löschen
       </div>
 
+      <!-- FilterIcon -->
+
       <TheIcons
         @click="modalStore.showModal = !modalStore.showModal"
         icon="bi bi-funnel"
         class="text-2xl border border-stone-500 p-2 h-full px-3 rounded-lg text-stone-600 hover:bg-stone-300 hover:text-white hover:duration-200 not-focus:duration-200"
       />
+      <!-- RefreshIcon -->
+
       <TheIcons
         @click="refreshData()"
         icon="bi bi-arrow-clockwise"
         class="text-2xl border border-stone-500 p-2 h-full px-3 rounded-lg text-stone-600 hover:bg-stone-300 hover:text-white hover:duration-200 not-focus:duration-200"
       />
     </div>
+
+    <!-- NORMAL Table -->
+
     <table v-if="tableStore.showTable === 'normalTable'" class="border border-separate w-full">
       <thead>
         <tr class="text-white bg-stone-500 w-full">
@@ -203,6 +181,8 @@ const showSearchBarDropdown = ref(false)
         </tr>
       </tbody>
     </table>
+
+    <!-- KOMPONENTE Table -->
 
     <table v-if="tableStore.showTable === 'komponentenTable'" class="border border-separate w-full">
       <caption class="caption-bottom">
