+++ src/views/AnwenderSettingsView.vue
@@ -1,5 +1,5 @@
 <script setup>
-import PaySite from '@/components/PaySite.vue'
+import PaySite from '@/components/site_overlay/PaySite.vue'
 </script>
 <template>
   <PaySite><div class="font-bold">"Anwender Settings"</div></PaySite>
