+++ src/views/GuideView.vue
@@ -1,5 +1,5 @@
 <script setup>
-import ConstructionSite from '@/components/ConstructionSite.vue'
+import ConstructionSite from '@/components/site_overlay/ConstructionSite.vue'
 </script>
 <template>
   <ConstructionSite><div class="font-bold">"Guide"</div></ConstructionSite>
